MAINT: Split pdf module
